add register, auth, login, logout...

+++ src/services/auth.js
@@ -15,8 +15,8 @@ export const registerUser = async (payload) => {
   const encryptedPassword = await bcrypt.hash(payload.password, 10);
 
   return await UsersCollection.create({
-    password: encryptedPassword,
     ...payload,
+    password: encryptedPassword,
   });
 };
 
@@ -26,13 +26,10 @@ export const loginUser = async (payload) => {
     throw createHttpError(404, 'User not found');
   }
   const isEqual = await bcrypt.compare(payload.password, user.password);
-  console.log('🚀 ~ loginUser ~ isEqual:', isEqual);
-  console.log('🚀 ~ loginUser ~ payload.password:', payload.password);
-  console.log('🚀 ~ loginUser ~ user.password:', user.password);
 
-  // if (!isEqual) {
-  //   throw createHttpError(401, 'Unauthorized');
-  // }
+  if (!isEqual) {
+    throw createHttpError(401, 'Unauthorized');
+  }
 
   await SessionsCollection.deleteOne({ userId: user._id });
 
@@ -48,43 +45,43 @@ export const loginUser = async (payload) => {
   });
 };
 
-// export const logoutUser = async (sessionId) => {
-//   await SessionsCollection.deleteOne({ _id: sessionId });
-// };
+const createSession = () => {
+  const accessToken = randomBytes(30).toString('base64');
+  const refreshToken = randomBytes(30).toString('base64');
 
-// const createSession = () => {
-//   const accessToken = randomBytes(30).toString('base64');
-//   const refreshToken = randomBytes(30).toString('base64');
+  return {
+    accessToken,
+    refreshToken,
+    accessTokenValidUntil: new Date(Date.now() + FIFTEEN_MINUTES),
+    refreshTokenValidUntil: new Date(Date.now() + ONE_DAY),
+  };
+};
 
-//   return {
-//     accessToken,
-//     refreshToken,
-//     accessTokenValidUntil: new Date(Date.now() + FIFTEEN_MINUTES),
-//     refreshTokenValidUntil: new Date(Date.now() + ONE_DAY),
-//   };
-// };
+export const refreshUsersSession = async ({ sessionId, refreshToken }) => {
+  const session = await SessionsCollection.findOne({
+    _id: sessionId,
+    refreshToken,
+  });
 
-// export const refreshUsersSession = async ({ sessionId, refreshToken }) => {
-//   const session = await SessionsCollection.findOne({
-//     _id: sessionId,
-//     refreshToken,
-//   });
+  if (!session) {
+    throw createHttpError(401, 'Session not found');
+  }
 
-//   if (!session) {
-//     throw createHttpError(401, 'Session not found');
-//   }
+  const isSessionTokenExpired =
+    new Date() > new Date(session.refreshTokenValidUntil);
 
-//   const isSessionTokenExpired =
-//     new Date() > new Date(session.refreshTokenValidUntil);
+  if (isSessionTokenExpired) {
+    throw createHttpError(401, 'Session token expired');
+  }
 
-//   if (isSessionTokenExpired) {
-//     throw createHttpError(401, 'Session token expired');
-//   }
+  const newSession = createSession();
 
-//   const newSession = createSession();
+  return await SessionsCollection.create({
+    userId: session.userId,
+    ...newSession,
+  });
+};
 
-//   return await SessionsCollection.create({
-//     userId: session.userId,
-//     ...newSession,
-//   });
-// };
+export const logoutUser = async (sessionId) => {
+  await SessionsCollection.deleteOne({ _id: sessionId });
+};
